Emulate bf16 via float32 output on sm75

+++ challenges/challenge_a_nf4.py
@@ -337,8 +337,8 @@ def _your_dequantize_nf4(
     out_dtype = _OUT_DTYPE_MAP[dtype]
     if dtype == torch.bfloat16 and major_version < 8:
         emulate_bf16 = True
-        out_storage_dtype = torch.float16
-        out_dtype = tl.float16
+        out_storage_dtype = torch.float32
+        out_dtype = tl.float32
 
     out_flat = torch.empty(n_weights, device=device, dtype=out_storage_dtype)
